requirements.txt
Flask
python-dotenv
requests
schedule
playwright
werkzeug==2.0.0
